requirements.txt
pyautogen>=0.10.0
autogen-ext>=0.1.0
tiktoken>=0.5.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
anthropic>=0.25.0
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
# pandas>=2.2.0  # Windows 컴파일 문제로 제거
numpy>=1.26.0
# lxml==4.9.3    # Windows 컴파일 문제로 제거
webdriver-manager==4.0.1 